move to tsx
